Add files via upload
The project objectives involve using Natural Language Processing (NLP) techniques to establish a harmonized and consistent terminology for the "CaseHistologicalDiagnosis" column in the dataset. The NLP approach will utilize tools such as the Unified Medical Language System (UMLS) and the MetaMap software to match diagnostic terms, handle synonyms, and ensure consistent coding. Additionally, the ConText tool will be employed to filter out negated concepts within the diagnoses.
The project requires expertise in Python Programming, as the NLP tools are developed in Python. Skills in managing data using SQL or NoSQL databases are also valuable for handling the dataset. The deliverables of the project include the source code for the NLP pipeline, a structured spreadsheet of canine diagnostic information with unified terminology, and preliminary univariate and multivariate logistic regression analyses of the processed data.